fix Issue #5, kd tree becomes unbalanced.
The problem was in the Memorizer.memorize method. When two series have the same first and second states but a different end state, we perform a weighted average. The problem was that because we were storing be reference to the original clusters, it was changing the values of the cluster arrays. This was breaking the kd tree. The series first, second and end states are now stored by value, fixing the issue.