Add secondary minheap sorting based on lvl property.
The lvl property allows for minheaps to be sorted based on temporal distance from their start state as a secondary property to sd. This should allow agents to act in a more intelligent manner in prioritizing behaviors, i.e. if two behaviors have the same sd value, choose the one that offers gratification sooner.