memory cap, and greatly improve rag doll demo performance
By switching the rag doll drive calculator to do increment and decrement (up to 30), combined with reducing the number of pattern matchers to 500, has resulted in dramatically better ability to stay upright. Warrants further investigation.

Also added a cap to the number of temporal planner memories are stored, to keep the engine from grinding to a halt with a combination of large number of pattern matchers and memory height.