add memory chain copy functionality
Add the ability to copy memory chains from one cluster to another. This feature did not result in noticeable improvements to performance tests, so it’s currently disabled. I’m keeping the new logic in the code base anyway, because I think it’s going to prove useful when modeling things like taxonomies.