Keep IHTAI from getting stuck in non-ideal loops.
This checkin should prevent IHTAI from ever getting stuck in a non-ideal loop. I’ve changed the logic in time series recording so that the second state now averages all moments with the same output array. Previously the comparison was based on input, output and drive state. This prevents situations where a certain set of stimuli occurs, paired with a low score, but IHTAI gets stuck in a loop accessing this score but experiencing a different memory. The score would never change, causing IHTAI to keep performing that action forever.